requirements.txt
# Deep Learning Frameworks
torch>=2.0.0
transformers>=4.30.0
openai>=1.0.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0
pytorch-lightning>=2.0.0

# Graph Neural Networks
torch-geometric>=2.3.0
dgl>=1.1.0

# Bayesian Networks
pgmpy>=0.1.23
pyro-ppl>=1.8.5
bayesnestor>=0.1.0

# Reinforcement Learning
gym>=0.26.0
stable-baselines3>=2.0.0

# Knowledge Graph Integration
rdflib>=6.3.2
SPARQLWrapper>=2.0.0

# NLP and Code Processing
tokenizers>=0.13.3
tree-sitter>=0.20.1
sentencepiece>=0.1.99

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.1

# API and Web
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
groq>=0.4.0
flask>=2.3.0

# Database
pymongo>=4.4.0
redis>=4.6.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# UI Framework
gradio>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

